+++ src/main/scala/org/globalforestwatch/features/SimpleFeatureId.scala
@@ -1,5 +1,5 @@
 package org.globalforestwatch.features
 
-case class SimpleFeatureId(featureId: Int) extends FeatureId {
+case class SimpleFeatureId(featureId: Long) extends FeatureId {
   override def toString: String = s"$featureId"
 }
